models-deployment/requirements.txt
Brotli==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.1.0
PySocks==1.7.1
PyYAML==6.0.1
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
exceptiongroup==1.2.0
fastapi==0.104.1
filelock==3.13.1
flask==3.0.0
fonttools==4.43.1
gmpy2==2.1.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
kiwisolver==1.4.5
matplotlib==3.8.0
mpmath==1.3.0
munkres==1.1.4
mysql-connector-python==8.2.0
nbconvert==7.6.0
networkx==3.2.1
numpy==1.26.0
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.2
patsy==0.5.3
pip==23.3.1
protobuf==4.21.12
psutil==5.9.5
py-cpuinfo==9.0.0
pydantic==2.0.3
pydantic-core==2.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
statsmodels==0.14.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.66.1
typing-extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.203
unicodedata2==15.1.0
urllib3==2.0.7
uvicorn==0.24.0.post1
websockets==12.0
werkzeug==3.0.1
wheel==0.41.3
xgboost==2.0.1
zipp==3.17.0